Extract amount in module
