+++ src/chain/bitcoin/mod.rs
@@ -3,7 +3,6 @@
 use bitcoin::hashes::sha256d::Hash as Sha256dHash;
 use bitcoin::secp256k1::Secp256k1;
 use bitcoin::secp256k1::Signature;
-use bitcoin::util::amount;
 use bitcoin::util::bip32::{DerivationPath, ExtendedPrivKey};
 use bitcoin::util::key::{PrivateKey, PublicKey};
 use bitcoin::util::psbt::PartiallySignedTransaction;
@@ -15,6 +14,7 @@ use crate::consensus::{self, AsCanonicalBytes, Decodable, Encodable};
 use crate::crypto::{self, Keys, SharedKeyId, SharedPrivateKeys, Signatures};
 use crate::role::Arbitrating;
 
+use amount::Amount;
 use transaction::{Buy, Cancel, Funding, Lock, Punish, Refund, Tx};
 
 use std::fmt::Debug;
@@ -24,6 +24,7 @@ use std::str::FromStr;
 pub mod fee;
 pub mod tasks;
 pub mod transaction;
+pub mod amount;
 
 #[derive(Clone, Debug, Copy, Eq, PartialEq)]
 pub struct Bitcoin;
@@ -57,53 +58,6 @@ impl Asset for Bitcoin {
 
     fn to_u32(&self) -> u32 {
         0x80000000
-    }
-}
-
-impl FromStr for Amount {
-    type Err = consensus::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let x = s
-            .parse::<u64>()
-            .map_err(|_| consensus::Error::ParseFailed("Failed to parse amount"))?;
-        Ok(Self(amount::Amount::from_sat(x)))
-    }
-}
-
-/// Bitcoin amount wrapper
-#[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Eq, StrictDecode, StrictEncode)]
-pub struct Amount(amount::Amount);
-
-impl Amount {
-    pub fn as_sat(&self) -> u64 {
-        self.0.as_sat()
-    }
-
-    pub fn from_sat(sat: u64) -> Self {
-        Self(amount::Amount::from_sat(sat))
-    }
-
-    pub fn checked_mul(&self, other: u64) -> Option<Self> {
-        Some(Self(self.0.checked_mul(other)?))
-    }
-
-    pub fn checked_sub(&self, other: Self) -> Option<Self> {
-        Some(Self(self.0.checked_sub(other.0)?))
-    }
-}
-
-impl Encodable for Amount {
-    fn consensus_encode<W: io::Write>(&self, writer: &mut W) -> Result<usize, io::Error> {
-        bitcoin::consensus::encode::Encodable::consensus_encode(&self.as_sat(), writer)
-    }
-}
-
-impl Decodable for Amount {
-    fn consensus_decode<D: io::Read>(d: &mut D) -> Result<Self, consensus::Error> {
-        let sats: u64 = bitcoin::consensus::encode::Decodable::consensus_decode(d)
-            .map_err(|_| consensus::Error::ParseFailed("Bitcoin amount parsing failed"))?;
-        Ok(Amount::from_sat(sats))
     }
 }
